refactor(theme): extract component family builders from hpe
